requirements.txt
requests~=2.28.0
schedule~=1.1.0
beautifulsoup4~=4.11.1
win10toast~=0.9
pystray~=0.19.4
Pillow~=9.2.0
elevate~=0.1.3